+++ lib/pages/login/component/login_header.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+
+class LoginHeader extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          "GYMMING",
+          style: TextStyle(
+            fontSize: 24,
+            fontWeight: FontWeight.bold,
+            fontFamily: 'Montserrat',
+            color: Colors.white,
+          ),
+        ),
+        SizedBox(
+          height: 2,
+        ),
+        Text(
+          "맞춤형 PT 일정을\n계획해보세요.",
+          style: TextStyle(
+            fontSize: 24,
+            fontWeight: FontWeight.w500,
+            color: Colors.white,
+          ),
+        ),
+      ],
+    );
+  }
+}
